WM: SharedContainerUtils split from SharedUtils
